+++ lib/mutant/meta/example/dsl.rb
@@ -80,14 +80,6 @@ module Mutant
           mutation('self')
         end
 
-        # Add regexp mutations
-        #
-        # @return [undefined]
-        def regexp_mutations
-          mutation('//')
-          mutation('/nomatch\A/')
-        end
-
         # Helper method to coerce input to node
         #
         # @param [String,Parser::AST::Node] input