+++ Changelog.md
@@ -2,6 +2,7 @@
 
 * New license.
 * Fix mutations to void value expressions to not be reported as integration error.
+* Remove regexp body mutations
 
 # v0.8.25 2018-12-31